login-form-and-quiz-app-project
This is a signup and login system where you can create multiple accounts. After creating an account, you are able to take a simple JavaScript quiz. Additionally, there is a logout button that will log you out from the quiz page when clicked. The quiz page is amazing: if you click on a wrong answer, it instantly turns red, and if you click on the right answer, the area turns green. This is my favorite project. I created many functions and learned a lot from it. Most of the form and quiz page is made using JavaScript.